+++ CHANGELOG.md
@@ -1,3 +1,63 @@
+## 0.8.0 (2020-10-13)
+
+### Features
+
+-   Added support time filter of casts
+-   Added support for new data providers (NTNU auv and Sintef auv)
+
+### Bug fixes and improvements
+
+-   Minor bugfixes
+-   Update packages to fix security issues
+-   Update to typings
+
+## 0.7.1 (2020-10-05)
+
+### Bug fixes and improvements
+
+-   Minor bugfixes
+
+## 0.7.0 (2020-09-22)
+
+### Features
+
+-   Added support for marine regions
+-   Refactor casts methods and types
+-   MRGID uses correct polygon instead of bounding box
+
+### Bug fixes
+
+-   Minor bugfixes
+
+## 0.6.4 (2020-09-11)
+
+### Features
+
+-   Added sequence metadata
+
+### Bug fixes
+
+-   Minor bugfixes
+
+## 0.6.2 (2020-09-09)
+
+### Features
+
+-   Added sequence metadata
+-   Added cast source file download
+
+### Bug fixes
+
+-   Minor bugfixes
+
+## 0.6.1 (2020-09-03)
+
+### Features
+
+### Bug fixes
+
+-   Minor bugfixes
+
 ## 0.3.1 (2020-03-24)
 
 ### Bug fixes